backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
pydantic[email]==2.9.2
aiofiles==24.1.0
openpyxl==3.1.5
opencv-python-headless>=4.8.0
python-dotenv>=1.0.0
